Modularize editor Markdown dialect
Separate schema, Markdown parsing, and serialization behind the existing conversion facade while preserving the public API and document behavior.
